src/auth/requirements.txt
alembic==1.7.5
aniso8601==9.0.1
arrow==1.2.1
attrs==21.4.0
Babel==2.9.1
binaryornot==0.4.4
blinker==1.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
cookiecutter==1.7.3
Deprecated==1.2.13
flasgger==0.9.5
Flask==2.0.2
Flask-BabelEx==0.9.4
flask-buzz==2.0.1
Flask-JWT-Extended==4.3.1
Flask-Limiter==2.0.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Pydantic==0.9.0
Flask-RESTful==0.3.9
Flask-Security==3.0.0
Flask-WTF==1.0.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jinja2-time==0.2.0
jsonschema==4.3.3
limits==2.2.0
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
mistune==2.0.1
orjson==3.6.5
packaging==21.3
passlib==1.7.4
pendulum==2.1.2
pluggy==1.0.0
poyo==0.5.0
psycopg2-binary==2.9.2
py==1.11.0
py-buzz==2.1.3
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
redis==4.1.0
requests==2.26.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.28
text-unidecode==1.3
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.0
zope.event==4.5.0
zope.interface==5.4.0
python-dotenv
flask_opentracing
jaeger_client
Authlib
pika